moved and refactores Blk::propagate_input_expressions(), Blk::merge_def_assignments_to_same_var(). Remove mocking functions of expression::builder.
