+++ src/cwe_checker_lib/src/intermediate_representation/blk.rs
@@ -1,9 +1,6 @@
 use super::*;
 use crate::utils::log::LogMessage;
-use std::{
-    collections::{HashMap, HashSet},
-    fmt,
-};
+use std::{collections::HashSet, fmt};
 
 /// A basic block is a sequence of `Def` instructions followed by up to two `Jmp` instructions.
 ///
@@ -70,131 +67,6 @@ impl Term<Blk> {
             Err(logs)
         }
     }
-
-    /// Wherever possible, substitute input variables of expressions
-    /// with the input expression that defines the input variable.
-    ///
-    /// Note that substitution is only possible
-    /// if the input variables of the input expression itself did not change since the definition of said variable.
-    ///
-    /// The expression propagation is used in [`expression_propagation` normalization pass](crate::analysis::expression_propagation)
-    /// to further simplify the generated expressions using a fixpoint algorithm
-    /// and allows more dead stores to be removed during [dead variable elimination](crate::analysis::dead_variable_elimination).
-    pub fn propagate_input_expressions(
-        &mut self,
-        apriori_insertable_expressions: Option<HashMap<Variable, Expression>>,
-    ) {
-        let mut insertable_expressions = HashMap::new();
-        if let Some(insertables) = apriori_insertable_expressions {
-            insertable_expressions = insertables;
-        }
-        for def in self.term.defs.iter_mut() {
-            match &mut def.term {
-                Def::Assign {
-                    var,
-                    value: expression,
-                } => {
-                    // insert known input expressions
-                    for (input_var, input_expr) in insertable_expressions.iter() {
-                        expression.substitute_input_var(input_var, input_expr);
-                    }
-                    // expressions dependent on the assigned variable are no longer insertable
-                    insertable_expressions.retain(|input_var, input_expr| {
-                        input_var != var && !input_expr.input_vars().into_iter().any(|x| x == var)
-                    });
-                    // If the value of the assigned variable does not depend on the former value of the variable,
-                    // then it is insertable for future expressions.
-                    if !expression.input_vars().into_iter().any(|x| x == var) {
-                        insertable_expressions.insert(var.clone(), expression.clone());
-                    }
-                }
-                Def::Load {
-                    var,
-                    address: expression,
-                } => {
-                    // insert known input expressions
-                    for (input_var, input_expr) in insertable_expressions.iter() {
-                        expression.substitute_input_var(input_var, input_expr);
-                    }
-                    // expressions dependent on the assigned variable are no longer insertable
-                    insertable_expressions.retain(|input_var, input_expr| {
-                        input_var != var && !input_expr.input_vars().into_iter().any(|x| x == var)
-                    });
-                }
-                Def::Store { address, value } => {
-                    // insert known input expressions
-                    for (input_var, input_expr) in insertable_expressions.iter() {
-                        address.substitute_input_var(input_var, input_expr);
-                        value.substitute_input_var(input_var, input_expr);
-                    }
-                }
-            }
-        }
-        for jump in self.term.jmps.iter_mut() {
-            match &mut jump.term {
-                Jmp::Branch(_) | Jmp::Call { .. } | Jmp::CallOther { .. } => (),
-                Jmp::BranchInd(expr)
-                | Jmp::CBranch {
-                    condition: expr, ..
-                }
-                | Jmp::CallInd { target: expr, .. }
-                | Jmp::Return(expr) => {
-                    // insert known input expressions
-                    for (input_var, input_expr) in insertable_expressions.iter() {
-                        expr.substitute_input_var(input_var, input_expr);
-                    }
-                }
-            }
-        }
-    }
-
-    /// Merge subsequent assignments to the same variable to a single assignment to that variable.
-    ///
-    /// The value expressions of merged assignments can often be simplified later on
-    /// in the corresponding [`Project` normalization pass](Project::normalize).
-    pub fn merge_def_assignments_to_same_var(&mut self) {
-        let mut new_defs = Vec::new();
-        let mut last_def_opt = None;
-        for def in self.term.defs.iter() {
-            if let Def::Assign {
-                var: current_var, ..
-            } = &def.term
-            {
-                if let Some(Term {
-                    term:
-                        Def::Assign {
-                            var: last_var,
-                            value: last_value,
-                        },
-                    ..
-                }) = &last_def_opt
-                {
-                    if current_var == last_var {
-                        let mut substituted_def = def.clone();
-                        substituted_def.substitute_input_var(last_var, last_value);
-                        last_def_opt = Some(substituted_def);
-                    } else {
-                        new_defs.push(last_def_opt.unwrap());
-                        last_def_opt = Some(def.clone());
-                    }
-                } else if last_def_opt.is_some() {
-                    panic!(); // Only assign-defs should be saved in last_def.
-                } else {
-                    last_def_opt = Some(def.clone());
-                }
-            } else {
-                if let Some(last_def) = last_def_opt {
-                    new_defs.push(last_def);
-                }
-                new_defs.push(def.clone());
-                last_def_opt = None;
-            }
-        }
-        if let Some(last_def) = last_def_opt {
-            new_defs.push(last_def);
-        }
-        self.term.defs = new_defs;
-    }
 }
 
 impl fmt::Display for Blk {
@@ -212,7 +84,6 @@ impl fmt::Display for Blk {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::intermediate_representation::{Def, Expression, Variable};
 
     impl Blk {
         /// Creates empty block with tid "block".
@@ -237,72 +108,5 @@ mod tests {
                 },
             }
         }
-    }
-
-    #[test]
-    fn expression_propagation() {
-        use crate::intermediate_representation::UnOpType;
-        let defs = vec![
-            Def::assign(
-                "tid_1",
-                Variable::mock("X", 8),
-                Expression::var("Y", 8).un_op(UnOpType::IntNegate),
-            ),
-            Def::assign(
-                "tid_2",
-                Variable::mock("Y", 8),
-                Expression::var("X", 8).plus(Expression::var("Y", 8)),
-            ),
-            Def::assign(
-                "tid_3",
-                Variable::mock("X", 8),
-                Expression::var("X", 8).un_op(UnOpType::IntNegate),
-            ),
-            Def::assign(
-                "tid_4",
-                Variable::mock("Y", 8),
-                Expression::var("Y", 8).un_op(UnOpType::IntNegate),
-            ),
-            Def::assign(
-                "tid_5",
-                Variable::mock("Y", 8),
-                Expression::var("X", 8).plus(Expression::var("Y", 8)),
-            ),
-        ];
-        let mut block = Term {
-            tid: Tid::new("block"),
-            term: Blk {
-                defs,
-                jmps: Vec::new(),
-                indirect_jmp_targets: Vec::new(),
-            },
-        };
-        block.merge_def_assignments_to_same_var();
-        block.propagate_input_expressions(None);
-        let result_defs = vec![
-            Def::assign(
-                "tid_1",
-                Variable::mock("X", 8),
-                Expression::var("Y", 8).un_op(UnOpType::IntNegate),
-            ),
-            Def::assign(
-                "tid_2",
-                Variable::mock("Y", 8),
-                Expression::var("Y", 8)
-                    .un_op(UnOpType::IntNegate)
-                    .plus(Expression::var("Y", 8)),
-            ),
-            Def::assign(
-                "tid_3",
-                Variable::mock("X", 8),
-                Expression::var("X", 8).un_op(UnOpType::IntNegate),
-            ),
-            Def::assign(
-                "tid_5",
-                Variable::mock("Y", 8),
-                Expression::var("X", 8).plus(Expression::var("Y", 8).un_op(UnOpType::IntNegate)),
-            ),
-        ];
-        assert_eq!(block.term.defs, result_defs);
     }
 }
