+++ src/cwe_checker_lib/src/intermediate_representation/expression/builder.rs
@@ -1,41 +1,9 @@
 #[cfg(test)]
-use apint::ApInt;
-
-#[cfg(test)]
-use super::{CastOpType, Variable};
-
+use super::CastOpType;
 use super::*;
 
 /// ## Helper functions for building expressions
 impl Expression {
-    /// Shortcut for creating a constant expression from an i64 value
-    #[cfg(test)]
-    pub fn const_from_i64(value: i64) -> Expression {
-        Expression::Const(Bitvector::from_i64(value))
-    }
-
-    /// Shortcut for creating a constant expression from an i32 value
-    #[cfg(test)]
-    pub fn const_from_i32(value: i32) -> Expression {
-        Expression::Const(Bitvector::from_i32(value))
-    }
-
-    /// Shortcut for creating a constant expression from an apint value (e.g. copy of global address)
-    #[cfg(test)]
-    pub fn const_from_apint(value: ApInt) -> Expression {
-        Expression::Const(value)
-    }
-
-    /// Shortcut for creating a variable expression
-    #[cfg(test)]
-    pub fn var(name: impl ToString, size_in_bytes: impl Into<ByteSize>) -> Expression {
-        Expression::Var(Variable {
-            name: name.to_string(),
-            size: size_in_bytes.into(),
-            is_temp: false,
-        })
-    }
-
     /// Shortcut for creating a cast expression
     #[cfg(test)]
     pub fn cast(self, op: CastOpType) -> Expression {
@@ -74,16 +42,6 @@ impl Expression {
         }
     }
 
-    /// Shortcut for creating an `IntSub`-expression
-    #[cfg(test)]
-    pub fn minus(self, rhs: Expression) -> Expression {
-        Expression::BinOp {
-            lhs: Box::new(self),
-            op: BinOpType::IntSub,
-            rhs: Box::new(rhs),
-        }
-    }
-
     /// Construct an expression that adds a constant value to the given expression.
     ///
     /// The bytesize of the value is automatically adjusted to the bytesize of the given expression.
@@ -99,20 +57,5 @@ impl Expression {
             _ => (),
         }
         self.plus(Expression::Const(value))
-    }
-
-    /// Construct an expression that subtracts a constant value from the given expression.
-    ///
-    /// The bytesize of the value is automatically adjusted to the bytesize of the given expression.
-    #[cfg(test)]
-    pub fn minus_const(self, value: i64) -> Expression {
-        let bytesize = self.bytesize();
-        let mut value = Bitvector::from_i64(value);
-        match u64::from(bytesize) {
-            size if size > 8 => value.sign_extend(bytesize).unwrap(),
-            size if size < 8 => value.truncate(bytesize).unwrap(),
-            _ => (),
-        }
-        self.minus(Expression::Const(value))
     }
 }
